fix(client/newsservices):add-conditional-in-newsservices(wait-for-headers-before-axios-request)

+++ client/src/services/newsServices.js
@@ -7,51 +7,61 @@ const token = localStorage.getItem('token');
 const headers = { 'Authorization': `Bearer ${token}` }
 
 export const getAllNews = async () => {
-  try {
-    const response = await axios.get(URLAPI_NEWS, { headers });
-    return response.data.sort((a, b) => (a.date < b.date) ? 1 : -1);
-  } catch (error) {
-    console.error('Error reading news:', error);
-    throw error;
+  if (headers) {
+    try {
+      const response = await axios.get(URLAPI_NEWS, { headers });
+      return response.data.sort((a, b) => (a.date < b.date) ? 1 : -1);
+    } catch (error) {
+      console.error('Error reading news:', error);
+      throw error;
+    }
   }
 };
 
 export const getNewsById = async (id) => {
-  try {
-    const response = await axios.get(`${URLAPI_NEWS + id}`, { headers });
-    return response.data;
-  } catch (error) {
-    console.error(`Error reading news`, error);
-    throw error;
+  if (headers) {
+    try {
+      const response = await axios.get(`${URLAPI_NEWS + id}`, { headers });
+      return response.data;
+    } catch (error) {
+      console.error(`Error reading news`, error);
+      throw error;
+    }
   }
 };
 
 export const createNews = async (newsData) => {
-  try {
-    const response = await axios.post(`${URLAPI_NEWS}`, newsData, { headers });
-    return response.data;
-  } catch (error) {
-    console.error('Error creating news:', error);
-    throw error;
+  if (headers) {
+    try {
+      const response = await axios.post(`${URLAPI_NEWS}`, newsData, { headers });
+      return response.data;
+    } catch (error) {
+      console.error('Error creating news:', error);
+      throw error;
+    }
   }
 };
 
 export const deleteNews = async (id) => {
-  try {
-    const response = await axios.delete(`${URLAPI_NEWS}${id}`, { headers });
-    return response.data;
-  } catch (error) {
-    console.error('Error deleting news:', error);
-    throw error;
+  if (headers) {
+    try {
+      const response = await axios.delete(`${URLAPI_NEWS}${id}`, { headers });
+      return response.data;
+    } catch (error) {
+      console.error('Error deleting news:', error);
+      throw error;
+    }
   }
 }
 
 export const updateNews = async (id, newsData) => {
-  try {
-    const response = await axios.put(`${URLAPI_NEWS}${id}`, newsData, { headers });
-    return response.data;
-  } catch (error) {
-    console.error('Error updating news:', error);
-    throw error;
+  if (headers) {
+    try {
+      const response = await axios.put(`${URLAPI_NEWS}${id}`, newsData, { headers });
+      return response.data;
+    } catch (error) {
+      console.error('Error updating news:', error);
+      throw error;
+    }
   }
 };
